Clarify jdk21u replacement (#3479)
Clarifying that "jdk21u" refers to the string in the command, not the directory produced by the scripts. It would be very sad if someone removed the jdk21u directory which contains the custom installation of gcc produced by the scripts, which would take a long time to regenerate.

(This almost happened to us when following the instructions due to the unclear wording.)